Create a new module to store all box/unbox routines.
